pyproject.toml
[project]
name = "alchemyql"
version = "0.0.1"
description = "Lightweight Graph QL engine powered by SQL Alchemy"
readme = "README.md"
authors = [
    { name = "Nicholas Williams", email = "noreply@users.noreply.github.com" }
]
license = "MIT"
license-files = ["LICENSE"]
keywords = ["graphql", "sqlalchemy", "database", "graphql-engine"]
homepage = "https://github.com/nicholasfelixwilliams/alchemyql"
repository = "https://github.com/nicholasfelixwilliams/alchemyql"
requires-python = ">=3.12"
dependencies = [
    "graphql-core>=3.2.7",
    "sqlalchemy>=2.0.44",
]

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "coverage>=7.12.0",
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
]
